+++ packages/substrate/src/wanxiang_substrate/evolution/promotion/__init__.py
@@ -1,0 +1,21 @@
+"""Promotion substrate (M30)."""
+
+from wanxiang_substrate.evolution.promotion.ladder import (
+    APPROVAL_LEVELS,
+    PROMOTION_LEVELS,
+    LevelRequirement,
+    PromotionEvidence,
+    PromotionLevel,
+    PromotionPolicy,
+    validate_promotion,
+)
+
+__all__ = [
+    "APPROVAL_LEVELS",
+    "PROMOTION_LEVELS",
+    "LevelRequirement",
+    "PromotionEvidence",
+    "PromotionLevel",
+    "PromotionPolicy",
+    "validate_promotion",
+]